apis/requirements.txt
flask==3.0.0
flask-pymongo==2.3.0
flask-pymongo==2.3.0
google-api-python-client==2.112.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
pymongo==4.6.0
Flask-RESTful==0.3.8
PyJWT==1.7.1
Flask-SQLAlchemy==2.4.1
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.9
pandas==2.1.3
openpyxl==3.1.2